backend/api/serializers/user_serialiser.py
from rest_framework import serializers
from backend.api.models.user import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'name', 'created_at', 'updated_at']