backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15.0
psycopg>=3.1.12
django-cors-headers>=4.3.0
gunicorn>=21.2.0